DAC demo for G4.
git-svn-id: svn://svn.code.sf.net/p/chibios/svn/trunk@13091 27425a3e-05d8-49a3-a47f-9c15f0e5edd8

+++ os/hal/ports/STM32/LLD/DACv1/hal_dac_lld.h
@@ -592,8 +592,9 @@ typedef enum {
   dacsample_t               init;                                           \
   /* DAC data holding register mode.*/                                      \
   dacdhrmode_t              datamode;                                       \
-  /* DAC control register.*/                                                \
-  uint16_t                  cr
+  /* DAC control register lower 16 bits.*/                                  \
+  uint32_t                  cr
+
 /**
  * @brief   Low level fields of the DAC group configuration structure.
  */
